#27 Função de utilidade e avaliação atualizadas com switch, para caso de IA x IA e no arquivo da IA
